+++ packages/docusaurus/eslint.config.js
@@ -1,8 +1,0 @@
-import config from "@incentro-ecx/eslint-config/docusaurus";
-
-export default [
-  ...config,
-  {
-    ignores: ["*.*"],
-  },
-];