+++ packages/docusaurus/src/index.ts
@@ -1,3 +1,0 @@
-export type { Config } from "./config.type.js";
-export { createConfig } from "./create-config.js";
-export type { PresetConfig } from "./preset-config.type.js";