+++ packages/docusaurus/src/preset-config.type.ts
@@ -1,1 +1,0 @@
-export type { Options as PresetConfig } from "@docusaurus/preset-classic";