+++ packages/eslint/eslint.config.js
@@ -1,13 +1,18 @@
+import { defineConfig } from "eslint/config";
 import config from "./dist/index.js";
 
-export default [
+export default defineConfig([
   ...config,
   {
+    name: "ignore-dist-and-root",
     ignores: ["dist/**", "*.*"],
   },
   {
+    name: "allow-default-exports",
+    files: ["**/*.js", "**/*.ts"],
     rules: {
-      'unicorn/expiring-todo-comments': 'off',
+      // It's common for ESLint configs to be the default export
+      "import/no-default-export": "off",
     },
-  }
-];
+  },
+]);