+++ packages/eslint/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @incentro-ecx/eslint-config
+# @incentro-ic/config-eslint
 
 ## 1.0.3
 
@@ -41,3 +41,103 @@
 ### Minor Changes
 
 - Improve import plugin compatibility with flat configuration for ESLint
+
+# Changelog from @micro.nova/eslint
+
+The original version of this package (which this is a fork off) was merged into this repository and released as version 2.0.0. Below is the changelog of all the versions from the original config:
+
+## 0.5.1
+
+### Patch Changes
+
+- Add base config to React Router ESLint config. ([#28](https://github.com/incentro-ecx/micro.nova/pull/28))
+
+## 0.5.0
+
+### Minor Changes
+
+- Add `@micro.nova/eslint/react-router` config for [React Router](https://reactrouter.com) projects. ([#27](https://github.com/incentro-ecx/micro.nova/pull/27))
+
+### Patch Changes
+
+- Update dependencies ([#27](https://github.com/incentro-ecx/micro.nova/pull/27))
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies. ([#23](https://github.com/incentro-ecx/micro.nova/pull/23))
+
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies. ([#20](https://github.com/incentro-ecx/micro.nova/pull/20))
+
+- Publish declaration maps and source code. ([#19](https://github.com/incentro-ecx/micro.nova/pull/19))
+
+## 0.4.0
+
+### Minor Changes
+
+- Add [spellchecker ESlint plugin](https://github.com/streetsidesoftware/cspell/tree/main/packages/cspell-eslint-plugin). ([#18](https://github.com/incentro-ecx/micro.nova/pull/18))
+
+- Add rules specific to test files using a [Vitest ESLint plugin](https://github.com/vitest-dev/eslint-plugin-vitest). ([#18](https://github.com/incentro-ecx/micro.nova/pull/18))
+
+### Patch Changes
+
+- Allow returning `void` in arrow function shorthands. ([#18](https://github.com/incentro-ecx/micro.nova/pull/18))
+
+- Allow use of `while(true)`. ([#18](https://github.com/incentro-ecx/micro.nova/pull/18))
+
+## 0.3.8
+
+### Patch Changes
+
+- Disable Typescript ESLint's `no-empty-object-type` rule for interface extensions ([#16](https://github.com/incentro-ecx/micro.nova/pull/16))
+
+## 0.3.2 - 0.3.7
+
+_Version 0.3.2 to 0.3.7 were published with errors and should not be used._
+
+## 0.3.1
+
+### Patch Changes
+
+- Re-order ESLint rules to overwrite recommended configs from plugins with our own overrides. ([#15](https://github.com/incentro-ecx/micro.nova/pull/15))
+
+## 0.3.0
+
+### Minor Changes
+
+- [BREAKING] Drop support for ESLint version 8 ([#13](https://github.com/incentro-ecx/micro.nova/pull/13))
+
+- Add names to custom configs to aid when using [ESLint Config Inspector](https://github.com/eslint/config-inspector) ([#13](https://github.com/incentro-ecx/micro.nova/pull/13))
+
+### Patch Changes
+
+- Update dependencies ([#13](https://github.com/incentro-ecx/micro.nova/pull/13))
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies ([#11](https://github.com/incentro-ecx/micro.nova/pull/11))
+
+## 0.2.1
+
+### Patch Changes
+
+- Turn off `unicorn/prefer-event-target` rule in favor of NodeJS's EventEmitter ([#10](https://github.com/incentro-ecx/micro.nova/pull/10))
+
+## 0.2.0
+
+### Minor Changes
+
+- Disabled `unicorn/no-useless-undefined` rule as it conflicts with Typescript settings around strict use of `undefined`. ([#9](https://github.com/incentro-ecx/micro.nova/pull/9))
+
+## 0.1.0
+
+### Minor Changes
+
+- Initial release ([#6](https://github.com/incentro-ecx/micro.nova/pull/6))